plugins/ocp: Move SMART cloud log methods to separate file for reusability
