+++ plugins/ocp/meson.build
@@ -2,5 +2,6 @@ sources += [
   'plugins/ocp/ocp-utils.c',
   'plugins/ocp/ocp-nvme.c',
   'plugins/ocp/ocp-clear-fw-update-history.c',
+  'plugins/ocp/ocp-smart-extended-log.c',
 ]
